Added Graphs for 6.2 using symbolic functions previously defined.
I added all of the 6.2 graphs, and corrected the mistake from the s3 file where angular velocity couldn't go bellow 0.  I also changed vx and vy to dx_du and dy_du respectively to make it easier to understand that it is seperate from the actual velocity that the neato is going. Also made minor changes to some formatting to ensure more consistency when viewed as a live script.